+++ src/main/java/org/aussiebox/circuit_core/helper/itemgroup/ItemGroupSupplier.java
@@ -1,0 +1,9 @@
+package org.aussiebox.circuit_core.helper.itemgroup;
+
+import net.minecraft.item.ItemGroup;
+import net.minecraft.util.Identifier;
+
+public interface ItemGroupSupplier {
+    Identifier getGroupId();
+    ItemGroup.Builder getGroupBuilder();
+}